copied code from dummy code to main code
