+++ templates/_base.html
@@ -5,6 +5,11 @@
             {% block title %} {% endblock title %}
         </title>
 	<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css">
+	<style>
+	*{
+		border-radius: 0 !important;
+	}
+	</style>
 	<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js"></script>
         {% block head %}
         {% endlbock head %}
